clean-up eyelink
