--- cardio_audio_sleep/_typing.py
@@ -1,7 +1,0 @@
-from abc import ABC
-
-
-class EYELink(ABC):  # noqa: B024
-    """Type hint for eye-link device."""
-
-    pass
